Fix linux-lab lint coverage
